+++ packages/bodiless-core/src/index.ts
@@ -27,10 +27,10 @@ import {
   withDefaultContent,
   withResetButton,
 } from './Contentful';
-import withEditButton, { useEditFormProps, createMenuOptionGroup } from './withEditButton';
+import withEditButton, { createMenuOptionGroup } from './withEditButton';
 import useContextMenuForm, { contextMenuForm, ContextMenuForm } from './contextMenuForm';
 import withCompoundForm, { useRegisterSnippet } from './withCompoundForm';
-import withEditFormSnippet from './withEditFormSnippet';
+import withEditFormSnippet, { useEditFormProps } from './withEditFormSnippet';
 import type { Options as EditFormSnippetOptions } from './withEditFormSnippet';
 import withData from './withData';
 import NodeProvider, { useNode, useNodeDataHandlers } from './NodeProvider';
